Jarvis_backup_20251008_1933/requirements.txt
openai>=1.51.0
python-dotenv>=1.0.1
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.26.4
pyttsx3>=2.90
